Chore #84: move files to new package
