+++ domain/my/src/main/java/com/hankki/domain/my/entity/response/MyJogboDetailEntity.kt
@@ -1,4 +1,4 @@
-package com.hankki.domain.my.entity
+package com.hankki.domain.my.entity.response
 
 data class MyJogboDetailEntity(
     val title: String,
