+++ domain/my/src/main/java/com/hankki/domain/my/entity/response/StoreEntity.kt
@@ -1,4 +1,4 @@
-package com.hankki.domain.my.entity
+package com.hankki.domain.my.entity.response
 
 data class StoreEntity(
     val category: String,
