+++ feature/my/src/main/java/com/hankki/feature/my/myjogbo/MyJogboRoute.kt
@@ -33,7 +33,7 @@ import com.hankki.core.designsystem.theme.Gray900
 import com.hankki.core.designsystem.theme.HankkiTheme
 import com.hankki.core.designsystem.theme.HankkijogboTheme
 import com.hankki.core.designsystem.theme.White
-import com.hankki.domain.my.entity.MyJogboEntity
+import com.hankki.domain.my.entity.response.MyJogboEntity
 import com.hankki.feature.my.R
 import com.hankki.feature.my.component.AddJogboItem
 import com.hankki.feature.my.component.DialogWithButton
