+++ feature/my/src/main/java/com/hankki/feature/my/myjogbo/model/MyJogboModel.kt
@@ -1,6 +1,6 @@
 package com.hankki.feature.my.myjogbo.model
 
-import com.hankki.domain.my.entity.MyJogboEntity
+import com.hankki.domain.my.entity.response.MyJogboEntity
 
 data class MyJogboModel(
     val jogboId: Long,
