+++ feature/my/src/main/java/com/hankki/feature/my/mypage/MyState.kt
@@ -1,6 +1,6 @@
 package com.hankki.feature.my.mypage
 
-import com.hankki.domain.my.entity.UserInformationEntity
+import com.hankki.domain.my.entity.response.UserInformationEntity
 
 data class MyState(
     val userState: UserInformationEntity = UserInformationEntity("", "")
