+++ feature/my/src/main/java/com/hankki/feature/my/mypage/MyViewModel.kt
@@ -2,7 +2,7 @@ package com.hankki.feature.my.mypage
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hankki.domain.my.entity.UserInformationEntity
+import com.hankki.domain.my.entity.response.UserInformationEntity
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
