+++ feature/my/src/main/java/com/hankki/feature/my/mystore/MyStoreRoute.kt
@@ -27,7 +27,7 @@ import com.hankki.core.designsystem.theme.Gray900
 import com.hankki.core.designsystem.theme.HankkiTheme
 import com.hankki.core.designsystem.theme.HankkijogboTheme
 import com.hankki.core.designsystem.theme.White
-import com.hankki.domain.my.entity.StoreEntity
+import com.hankki.domain.my.entity.response.StoreEntity
 import com.hankki.feature.my.R
 import com.hankki.feature.my.component.StoreItem
 import com.hankki.feature.my.mystore.model.MyStoreModel
